refactor: move status attribute class to the payment folder
